\name{syntheticTrx}
\alias{syntheticTrx}
\docType{data}
\title{
  Populus trichocarpa synthetic transcript annotation
}
\description{
  Annotation for P. trichocarpa retrieved as a gff file from Phytozome 10 
  (P. trichocarpa version 3) and post-processed to collapse transcript 
  information into a single representative transcript. This is a critical
  step for accurate RNA-Seq per gene counting.
}
\value{
  A \linkS4class{GRanges} object.
}
\seealso{
  The package vignette: RnaSeqTutorial.
}
\keyword{datasets}
