% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{Density plot}
\alias{Density plot}
\alias{densityPlot}
\alias{densityPlot,numeric,numeric-method}
\title{Density plot adapted from the LSD package}
\usage{
densityPlot(x, y, grid = 100, ncol = 30, nlevels = 10, ...)
}
\arguments{
\item{x}{a numeric vector for the x axis value}

\item{y}{a numeric vector for the y axis value}

\item{grid}{the size of the grid, default to 100}

\item{ncol}{the number of color to use, default to 30}

\item{nlevels}{the number of levels passed as argument to the 
\code{\link[graphics:contour]{contour}} function, defaults to 10}

\item{...}{additional argument passed to the 
\code{\link[graphics:image]{image}} function.}
}
\description{
Draws a topographic density plot.
}
\details{
This function was extracted from the LSD package and
was modified to suppress the change of margins
}
\examples{
\dontrun{
## TODO add an object after putting results in the data folder
## and use that for plotting the example
data(RST-DataFrame)
sel <- ! is.na(dataFrame$padj)
densityPlot(log10(dataFrame$baseMean),
            dataFrame$log2FoldChange)
}
}
\seealso{
\code{\link[graphics:contour]{contour}},
\code{\link[graphics:image]{image}} and
\code{\link[LSD:heatscatter]{heatscatter}}
}

