\name{Deprecated data}
\alias{Deprecated data}
\title{The following data has been deprecated:
\itemize{
\item annot.gff
}}
\description{
  \itemize{ \item The annot.gff file containing exon annotation 
from FlyBase v5.29 (http://flybase.org, Tweedie et al., 2009) has
been superseded by the Dmel-mRNA-exon-r5.52.gff3 file from
FlyBase v5.52 (April 2013) containing both mRNA and exon annotation
type.
  }
}

