% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{Package data accessor}
\alias{Package data accessor}
\alias{extdata}
\alias{extdata,missing-method}
\title{Data accessor for the RnaSeqTutorial package}
\usage{
extdata(x)
}
\arguments{
\item{x}{missing}
}
\value{
The file path to the RnaSeqTutorial package extdata folder
}
\description{
Simply lookup the extdata path.
}
\details{
This function simply lookup the extdata path in the installation directory
of the RnaSeqTutorial package
}
\examples{
  extdata()
  
}

