% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalGenomics.R
\name{Functional Genomics High Level}
\alias{Functional Genomics High Level}
\alias{readAbundance}
\alias{readAbundance,character-method}
\alias{nonExpressed}
\alias{nonExpressed,matrix-method}
\alias{rawDataMeanPlot}
\alias{rawDataMeanPlot,matrix-method}
\alias{rawDataSamplePlot}
\alias{rawDataSamplePlot,matrix-method}
\alias{createDESeqDataSet}
\alias{createDESeqDataSet,matrix,data.frame-method}
\alias{reportSizeFactors}
\alias{reportSizeFactors,DESeqDataSet-method}
\alias{transform}
\alias{transform,DESeqDataSet-method}
\alias{plotUnTransformed}
\alias{plotUnTransformed,DESeqDataSet-method}
\alias{plotPca,}
\alias{plotPca,matrix,data.frame-method}
\title{Functional Genomics High-Level functions}
\usage{
readAbundance(x, pattern = character(0), type = c("kallisto",
  "salmon"))
}
\arguments{
\item{x}{a character value (readAbundance), a DESeqDataSet (reportSizeFactor,
transform) or a matrix (all other functions)}

\item{pattern}{a character pattern to match the abundance files (readAbundance)}

\item{type}{a character value to define the type of quantification (readAbundance)}

\item{y}{a data.frame (createDESeqDataSet amd plotPca)}

\item{design}{a design expression (createDESeqDataSet)}
}
\value{
TODO FIXME
}
\description{
High level functions for the Functional Genomics course computer lab
}
\examples{
\dontrun{
  counts <- readAbundance(x="path-to-kallisto-output-dir")
  nonExpressed(counts)
  rawDataMeanPlot(counts)
  rawDataSamplePlot(counts)
}
}
