% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcFunction.R
\name{GC calculation}
\alias{GC calculation}
\alias{gcFunction}
\alias{gcFunction,DNAStringSet-method}
\title{GC calculation}
\usage{
gcFunction(x)
}
\arguments{
\item{x}{a DNAStringSet}
}
\value{
The percent GC of the sequences in \code{x}
}
\description{
Simple function to calculate the GC percentage of sequences
}
\details{
This function simply lookup the amount of G and C bp in DNAStringSet and
return the sequences GC proportion. The function was originally written
by Martin Morgan for the original version of that package.
}
\examples{
\dontrun{
  gcFunction(someDNAStringSet)
}
}
