% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMALSD.R
\name{Heatscatter based plot}
\alias{Heatscatter based plot}
\alias{plotMALSD}
\alias{plotMALSD,data.frame-method}
\title{Heatscatter based plot}
\usage{
plotMALSD(x, ylim, sign = 0.05, col = "forestgreen",
  linecol = "#00000080", xlab = "mean of normalized counts",
  ylab = expression(log[2] ~ fold ~ change), log = "x", cex = 0.45,
  ...)
}
\arguments{
\item{x}{a data.frame object, the result of the \code{DESeq} analysis. 
It must contain the following columns: \code{baseMean}, 
\code{log2FoldChange} and \code{padj}.}

\item{ylim}{the y axis limits}

\item{sign}{the significance threshold cutoff (on the adjusted p-values). 
Default to 5 percent}

\item{col}{the color to mark the points exceeding the significant threshold.
Default to 'forestgreen'}

\item{linecol}{the line color, default to"#00000080"}

\item{xlab}{the x axis label. Default to "mean of normalized counts".}

\item{ylab}{the y axis lable. Default to "dispersion" or 
"expression(log[2] ~ fold ~ change)" for the dispertion and MA plot, 
respectively.}

\item{log}{whether to report the axis on a log scale. default to "x" and 
"xy" for the MA and dispersion plot, respectively. See 
\code{\link[graphics:par]{par}} for more details.}

\item{cex}{the character extension, default to 0.45}

\item{...}{additional argument passed to the 
\code{\link[LSD:heatscatter]{heatscatter}} function.}
}
\description{
A based on the \code{\link[LSD:demotour]{LSD}} package
\code{\link[LSD:heatscatter]{heatscatter}} function. It is an
adaptation of the MA plot from the \code{DESeq} package 
\code{\link[DESeq:plotMA]{plotMA}} function.
}
\details{
The plot is a modification that integrate a third dimension as a colouring
shade from blue (sparse) to yellow (dense). Grey represent outlier data 
points. This is based on the \code{\link[LSD:heatscatter]{heatscatter}} 
function of the \code{\link[LSD:demotour]{LSD}} package.
}
\examples{
\dontrun{
## TODO add an object after putting results in the data folder
## and use that for plotting the example

data(RST-DataFrame)
plotMALSD(dataFrame)
}

}
\seealso{
\code{\link[LSD:demotour]{LSD}},
\code{\link[LSD:heatscatter]{heatscatter}},
\code{\link[DESeq:plotDispEsts]{plotDispEsts}},
\code{\link[DESeq:plotMA]{plotMA}}
}
