% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorial-package.R
\docType{package}
\name{RnaSeqTutorial package}
\alias{RnaSeqTutorial package}
\alias{RnaSeqTutorial package-package}
\alias{RnaSeqTutorial-package}
\title{Count summarization and normalization pipeline for Next Generation
Sequencing data.}
\description{
Offers functionalities to summarize read counts per feature of interest,
e.g. exons, transcripts, genes, etc.
Offers functionalities to normalize the summarized counts using 3rd party
packages like \code{\link[DESeq:newCountDataSet]{DESeq}}
or \code{\link[edgeR:DGEList]{edgeR}}.
}
\examples{
    \dontrun{
    library("RnaSeqTutorial")
  TODO put some examples there
    }

}
\author{
Nicolas Delhomme, Bastian Schiffthaler, Ismael Padioleau
}
\seealso{
\code{\link[easyRNASeq:easyRNASeq-package]{easyRNASeq}}
}
\keyword{package}

