\name{gAnnot}
\alias{gAnnot}
\docType{data}
\title{
  Drosophila melanogaster genic annotation retrieved from FlyBase v5.29 (June 2010)
}
\description{
  Annotation for D. melanogaster retrieved as a gff file from FlyBase (Tweedie et al.,2009) and post-processed. This file should not be used for analyses purposes.
}
\usage{	
	data(gAnnot)
}
\seealso{
  The package vignette: RNAseq.use.case.
}
\keyword{datasets}

