% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA.R
\docType{methods}
\name{MA plot}
\alias{MA plot}
\alias{plotMA}
\alias{plotMA,DataFrame-method}
\title{MA (median vs. average) plot}
\usage{
\S4method{plotMA}{DataFrame}(object, alpha = 0.001)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{DataFrame}} having at
least 3 columns named \code{baseMean}, \code{log2FoldChange} and \code{padj}}

\item{alpha}{the significance cutoff to be used when plotting, defaults to
10e-3.}
}
\value{
TRUE invisibly
}
\description{
Draws an MA plot, in the present case (the log2 fold changes vs. the log10 
of the mean)
}
\details{
This function has been devised to draw a series of MA plot from a 
\code{DESeq} result object.
}
\examples{
\dontrun{
## TODO add an object after putting results in the data folder
## and use that for plotting the example
data(RST-DataFrame)
volcanoPlot(dataFrame)
}

}
\seealso{
\code{\link[DESeq:newCountDataSet]{DESeq package}}
}
