% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{Volcano plot}
\alias{Volcano plot}
\alias{volcanoPlot}
\alias{volcanoPlot,DataFrame-method}
\title{Volcano plot}
\usage{
volcanoPlot(object, alpha = 0.001)
}
\arguments{
\item{object}{an object of class \code{\linkS4class{DataFrame}} having at
least 2 columns named \code{log2FoldChange} and \code{padj}}

\item{alpha}{the significance cutoff to be used when plotting, defaults to
10e-3.}
}
\description{
Draws a Volcano plot, i.e. the differential expression log 
odds vs. expression log fold change between condition.
}
\details{
The function has been devised to draw a MA plot from a \code{DESeq}
result object.
}
\examples{
\dontrun{
## TODO add an object after putting results in the data folder
## and use that for plotting the example
data(RST-DataFrame)
volcanoPlot(dataFrame)
}

}
\seealso{
\code{\link[DESeq:newCountDataSet]{DESeq package}}
}

